// frontendAng/src/app/model/goblet.ts
export class Goblet {
    _id:String;
    tQuestion7: String
    tAnswer7: String;
    tQuestion6: String;
    tAnswer6: String;
    tQuestion5: String;
    tAnswer5: String;
    tQuestion4: String;
    tAnswer4: String;
    tQuestion3: String;
    tAnswer3: String;
    tQuestion2: String;
    tAnswer2: String;
    tQuestion1: String;
    tAnswer1: String;
    bQuestion2: String;
    bAnswer2: String;
    bQuestion3: String;
    bAnswer3: String;
    bQuestion4: String;
    bAnswer4: String;
    bQuestion5: String;
    bAnswer5: String;
    bQuestion6: String;
    bAnswer6: String;
    bQuestion7: String;
    bAnswer7: String
}